Code cleanup; utilize Cog pattern; Added modules
