+++ modules/discord_utils.py
@@ -1,0 +1,11 @@
+
+async def send_multipart_msg(ctx, raw_msg):
+    msg_len = len(raw_msg)
+    # >>> chunks, chunk_size = len(x), len(x)/4
+    # >>> [ x[i:i+chunk_size] for i in range(0, chunks, chunk_size) ]
+    #num_msg_required = math.ceil(msg_len / 2000)
+    msg_chunk_size = 2000
+    # parts = [your_string[i:i + n] for i in range(0, len(your_string), n)]
+    msg_part_list = [raw_msg[i:i + msg_chunk_size] for i in range(0, msg_len, msg_chunk_size)]
+    for i in range(0, len(msg_part_list)):
+        await ctx.send(msg_part_list[i])
